requirements.txt
# Base
Django==4.0.4
django-allauth==0.50.0
django-filter==21.1

# REST
djangorestframework==3.13.1
dj-rest-auth==2.2.4
# django-cors-headers==3.7.0

# Other
prodict==0.8.18
# boto3

# Monitor
sentry-sdk==1.5.10

# Telegram bot
python-telegram-bot~=13.11
dynaconf~=3.1
